Update value of shortName field
To fix the problem that the UCR document in GH space points to 

https://www.w3.org/TR/dxwg-ucr/

instead of

https://www.w3.org/TR/dcat-ucr/

How this revision addresses the problem:

As explained in the relevant [ReSpec documentation page](https://github.com/w3c/respec/wiki/shortName), field `shortName` is used to determine the last part of specification URL in TR space. As the UCR URL in TR space is https://www.w3.org/TR/dcat-ucr/ , the value of field `shortName` should be `dcat-ucr`.