Edits wrt RPFDAPUB and RPFDPUB
Commented out RPFDPUB and kept only RPFDAPUB in the general profile part. Created a new requirements to reflect that the Google doc has one that is overlapping with that didn't have a github issue. Created a new github issue to it. Used very similar heading and abbreviations (RPFDAPUB-1 and RPFDAPUB-2) to reflect that all this is going to require some effort, and wrote notes to sum up the situation in both requirements.