fix issue https://github.com/w3c/dx-connegp/issues/40
the URL https://w3c.github.io/dxwg/connegp/ is used in some places
to refer to the editor's draft (I assume it was first located in this repo,
before being moved to its own repo).

This PR creates a "redirect" from the old URL to the new one,
in order to "fix" the broken links that still exist out there.